Enable Emmet in HTML+ERB files (#1)
Hi! As part of [this
issue](https://github.com/zed-extensions/ruby/issues/162) we would like
to rename `HTML/ERB` to `HTML+ERB` since it is more syntactically
correct to treat such language as ERB on top of HTML rather than HTML or
ERB.

To keep the user experience intact, we outlined the prerequisites in the
linked issue. This is the first PR that adds the `HTML+ERB` language
name to the list of enabled languages for the Emmet extension. We will
do the same for the Tailwind configuration in the Zed codebase. Once the
new versions of Emmet and Zed are released, we will merge the pull
request in the Ruby extension repository and release the updated
version. After that, we will remove the old `HTML/ERB` and `YAML/ERB`
languages. Let me know if that sounds good. Thanks!